Changed behavior of device discovery and connection.
Now device discovery is performed after a call to scanForNewDevices
and onDeviceListUpdated callback is always called when discovery is
completed. TransportManager does not automatically connects to all
discovered devices. Client must call connectDevice to mark device as
connected and disconnectDevice to mark device as disconnected.
When scanForNewDevices is not requested, discovery thread will only
perform service discovery for connected devices and will update connections
for these devices (i.e. if AppLink service is discovered at some channel
that is not yet connected it will initiate connection and if AppLink service
is no longer available on channel with existing connection this connection
will be terminated).