New logger settings.
Now there are 4 appenders:

1. ConsoleAppender. Logs only ERROR and FATAL messages to console using
   shorter conversion pattern (no filename, line number and function name).
2. DailyRollingFileAppender with output to AppLinkCore.log. All AppLinkCore
   messages are logged here.
3. DailyRollingFileAppender with output to AppLinkCoreWarn.log. Only
   WARN, ERROR and FATAL AppLinkCore messages are logged here.
4. DailyRollingFileAppender with output to TransportManager.log. All
   TransportManager messages are logged here.

All DailyRollingFileAppenders are set to DAILY, i.e. current date will be
appended to log file name. If log file with this name already exists it
will be moved to *.log.1, *.log.1 will be in turn moved to *.log.2 and so on.
Each run of appLinkCore will produce separate set of log files.