APPLINK-3551: Bluetooth transport: Attempt to fix a bug when sending a lot of data on Android prior to 4.2
The OS before 4.2 has an old Bluetooth stack, which behaves very odd in
some cases when sending a lot of data for a while. It's not reproducible
when the smartphone is connected to the debugger, probably because it
generates a lot of logs, and thus decreases the speed.

An ugly hack is to drop the speed artificially by sleeping for some time.
The magic number is discovered experimentally, and may not work for all
devices.