APPLINK-8424. Fix for incorrect inserting new group
Since we uses this ifunctional_group_id in many tables, we have to be sure
that id for certain group will be same in case when
we drop records from the table and add them again.
That's why we use hash as a primary key insted of
simple auto incremental index.