Fix putting NAVI non-media app in BACKGROUNG when activating a second MEDIA media app
According to the requirements (APPLINK-8531), SDL must support the apps
of the following AppHMIType to be in LIMITED at one and the same time:
MEDIA media, NAVIGATION non-media, COMMUNICATION non-media. But because
of a bug, when activating a MEDIA media app, the previously activated
NAVIGATION non-media app was put from LIMITED to BACKGROUND. It should
stay in LIMITED.
The fix is taken from CustomSDL/sdl_panasonic/develop(451042b4)
'APPLINK-13597 Buxfix in StateCtrl'.

Fixes: APPLINK-20258