Fix sending Policy Table snapshot in OnSystemRequest
Before this change OnSystemRequest notification with PT snapshot and an
update URL, sent by SDL to mobile when the mobile app registers, was
actually empty. The reason was that when sending the notification we
were trying to read PT data from a non existing file, while the actual
data was discarded.
This change fixes the problem by removing reading from file and using
the already built PT data in memory.
It also fixes sending the PTU url in the message (it was also missing).

Fixes: APPLINK-17788