Increase Slider mobile request timeout to receive HMI response
Until now we were setting generic command timeout for Slider request
equal to Slider's own timeout immediatelly after receiving the command.
Then we send the command to HMI and wait for HMI response. If we don't
do anything in HMI it sends TIMED_OUT response to SDL, but this response
comes AFTER the generic command timeout has expired (because it was
started a litle bit earlier). The generic timeout triggers GENERIC_ERROR
response and HMI's TIMED_OUT response is ignored.
The fix is to set generic command timeout bigger that slider timeout.

Fixes: APPLINK-20738