hotfix/SDL does not release memory after app unregistering
Registered after start connections in TransportAdapterImpl
can be unregistered when finalized using the RemoveFinalizedConnection
method from the TransportAdapter interface.

The call to the RemoveConnection method in the TransportManagerImpl
unregisters the connection calling the RemoveFinalizedConnection
using the supplied TransportAdapter instance. This solves the
observed memory leak of the inactive connections.

Issues: APPLINK-22910, APPLINK-23563