Fix_OnSystemRequestNotification_sent_to_Mobile
    The main reason for the HMI notification to not be sent to Mobile side
was the missing appID which was causing fail check in app.valid(). The appID
was missing, because it was not sent from HMI. In this case we loop over
the registered applications and take the first which device is allowed
by policy.[see APPLINK-12696].

- Add app_id handling in on_system_request_notification.cc
- Modify MOBILE_API.xml with the correct maxvalues for OnSystemRequest`s
offset and lenght parameters. Also remove whitespace at the eol.

Related: APPLINK-12369
	[https://adc.luxoft.com/jira/browse/APPLINK-12369]