Parse PT String for HTTP Request Body
The "body" portion of the httprequest needs to be a valid string. Before this fix, the policy table content of the request did not have properly escaped double quotes and was causing the JSON received by the mobile side to be considered invalid.

The new method "ParsePtString()" adds an escaped backslash during every occurance of a double quote or other escaped backslash. This method also removes any newlines located in the policy table string.