Fix_core_dump_when_app_is_in_LIMITED_and_ignition_off_is_sent
The reason for the coredump was a string destructor call after the
string was already destructed. After checking the address in memory
where we were dummped the involved strings were discovered. The
apperant reason for the double destruction in doubled linking of
the same string in static and dynamic library.

Related-Issue: APPLINK-22355