Fix "Content-Length" Spelling and Calculation
Content-Length was using an underscore instead of a Hyphen. Android was throwing an exception because there was no "Content-Length" key supplied in the request header.

Fixed the way the Content-Length is calculated. Android removes the extra backslashes before sending the request to sdl server so I have Content-Length calculated first before "ParsePTString()" is called. Then contentLength is sent to ParsePtString to be adjusted if characters are removed from the parsed policy table string (ie '\n' the newline character).

Tested on android and ios.