Fixes LAUNCH_APP flow
On invoking of SDL.ActivateApp SDL has to search through the registered
and to-be-regisetered apps in order to find necessary app to launch.
In case of registered app SDL has just to start it.
In case of to-be-registered app SDL has to send LAUNCH_APP request via
foregrounded SDL4.0-enabled application. If there is no such app - request
has to be sent via all registered SDL4.0-enabled application on the same
device as application requested to be launched.

Relates-to: APPLINK-23987