Check return values for profile file functions (CID 79994)
This commit addresses the Coverity defect indicating that the remove()
library function was unchecked. The first remove should no be there
since rename will replace and remove the file already. If rename does
not work, that file should also not be destroyed.

The second remove has its return value checked and logged if an error
exists.