Fix_secure_service_opening_without_validation
The problem which is occuring is that when we remove the server
cridentials from SDL and start secure service we succeed.
The reason for that is the VerifyPeer located in .ini file which by
default was set false. As result SSL_CTX_set_verify is called
without SSL_VERIFY_PEER, which is the reason server`s
cridentials to not be checked.

Related-issue: APPLINK-17757