Change_VerifyPeer_parameter_in_ini_file_to_true
By implementing this feature we fix secure service opening
without validation.The problem which is occuring is that
when we remove the server cridentials from SDL and start
secure service we succeed. The reason for that is the VerifyPeer
located in .ini file which by default was set false. As result
SSL_CTX_set_verify is called without SSL_VERIFY_PEER, which is
the reason server`s cridentials to not be checked.

Requirement: APPLINK-22607
Fix-issue: APPLINK-17757