Change SDL_SQLITE_DIR to SDL_SQLITE_BIN_DIR
There is a contradiction in the installation instructions for windows about using SDL_SQLITE_DIR environment variable.
For example in file make_sqlite_lib.cmd is assumed it points to the sqlite binary directory but in other places it is expected to point to sqlite source directory (FindSDLSqlite3.cmake).
That's why it is better to use different environment variables. Instructions must be updated too.