Fix DCHECK failure for policy retry sequence timer
The `join()` function fails when stopping the policy retry sequence
timer in the policy manager `RetrySequence` function.

Since the policy retry sequence timer calls the `RetrySequence` function
itself, there is no need to use `Stop()` to end the thread. `Stop()`
uses a join which requires the calling thread to be different than the
thread to be joined.

Since this is not the case, we can just return from the function which
will finish the thread without needing the `Stop()` function.