Update README to reflect run changes for 4.1.0
For the most recent master, smartDeviceLinkCore now needs to be run from build/bin to work correctly.  This pull request updates the README to reflect this.  Also, the keys do not need to be copied over to src/appMain since the executable can now be run from the build/bin directory.