Fix HMI resumption after activation
If an HMI application is activated to HMI level FULL, it could still be
resumed which causes the application to then receive HMI level NONE.

This is because the HMI level is set during the initial resumption
function using `SetupDefaultHMILevel` which currently only sets the HMI
level to NONE.

The HMI level should only be set if resumption is still needed.  This
occurs in the `StartAppHmiStateResumption` function. If the app is
activated, it is removed from the resumption queue in
`RemoveFromResumption`. This ensures that the HMI is not set during
resumption when the app is already active.

Therefore, there is no need to have `SetupDefaultHMILevel` in the
initial resume function since `SetupDefaultHMILevel` should only be
called when the app actually needs resumption.