Reorder register app response to mobile and hmi
If an HMI activates an application immediately after receving a
register app notification, then the mobile application could receive a
HMI FULL notification from the activation and then an HMI NONE
notification. The HMI NONE notification comes from the resumption
function following the HMI and mobile responses.

By sending the HMI notification last, we can ensure that resumption will
follow before the HMI sends an `ActivateApp`. With this commit, the
mobile application will first be notified of successful registration and
then the HMI will be notified.