Change deprecated readdir_r to readdir
`readdir_r` is deprecated as of glibc 2.24 in favor of `readdir`.  In
the use cases of `readdir_r`, it can safely be replaced with `readdir`.

POSIX specifies that "The pointer returned by readdir() points to data
which may be overwritten by another call to readdir() on the same
directory stream. This data is not overwritten by another call to
readdir() on a different directory stream".

Since the code does not reuse the directory stream in a way that could
cause threading issues, we can safely use `readdir`.  This prevents
possible issues with buffer overflows in `readdir_r` and simplifies the
code.