Remove Policy_lib_check_void() for activate app
In the event policies are disabled, this policy lib check void would return before sdl has the opportunity to tell the hmi to activate the app. When this function returns because of policy lib check void, an app could never be activated and no warnings or errors are reported.

This check of policies is not needed either since there are the appropriate checks in this function for if policies are enabled, and handles either setting appropriately.