Fix SDL BC.ActivateApp for NONE level
Fixed problem that SDL BC.ActivateApp wasn't sent in case of user replies
"NO" for data consent prompt.
The problem was in additional check that SDL BC.ActivateApp could be sent only if
HMI level was FULL. But in case of negative user reply HMI level must be NONE.
Therefore this check was deleted.

Related: APPLINK-30617