Fix sending URLs of unregister apps in GetURLs
Although there is a check for unregister apps, when the
URLs are gathered, their links appear in the response,
because the URL counting index increments independently
of the type of the application.

The URL counting index is decremented if current app is
not registered, so that its links are skipped.

Related-issues: APPLINK-30883