Fix sending same URL in OnSystemRequest retry
SDL sends same URL in every OnSystemRequest retry, because
for every message is used first url of the first application.

As a fix, for each message is used the next URL from the current
application URLs vector, starting from first valid application.
It there is no more URLs for this application, the function looks
for the next registered application (or default) and begin cycling
through its URLs.

Related-issues: APPLINK-30859