Fix OnSystemRequest cycling urls refactoring
Required refactoring of the functionality that is responsible
for sending different application URL on each OnSystemRequest.
The logic was placed in OnSnapshotCreated method. Now it is
separated on two methods - GetNextUpdateUrl - returns pair of
policy application id and url from the Endpoints vector that
potentially will be sent, IsUrlAppIdValid - checks if given
policy application id is assigned to a registered application
or it is the default id.