Fix wrong info parameter AudioPassThru
In case received from mobile side audioPassThruIcon does not exist in
the application storage folder SDL used to add string "Reference image(s)
not found" to the info parameter in the response. After clarification it tur
out that this info must be generated from the HMI, so that the member variab
that was used for storing the state of the icon is removed.