Fix omited URL in OnSystemRequest retry
The URL is skipped because of a new application is
registered and it is received the index of the default
section in the Endpoints vector. Fix compares the policy
application id of the application whose URL is about
to be sent from the Endpoints vector and from the member
variable in PolicyManager. If there is a difference,
the URL index is set to 0, so that no URL will be omited.