Fix ResetGlobalProperties help prompt parameter
Fixed ResetGlobalProperties help prompt request parameter reset to default
value. The root cause of this defect is that SDL does not get default help
prompt value from Application Manager settings and does not assign it to
app's help prompt. It just assigns an empty map.

In this commit:
- Added missed help_prompt() function to ApplicationManagerSettings interface
- Added related mock method to MockApplicationManagerSettings class
- Updated ResetGlobalPropertiesRequest::ResetHelpPromt function logic to work
as expected