Fix UnregisterAppInterface for not found app
If an app is not found in SDL but the mobile side
sends UnregisterAppInterface, SDL creates the mobile
request but never runs it because the app is not found.
Then the request times out and the response to the
mobile side is `GENERIC_ERROR`. Now command will be
ran and the proper response of `APPLICATION_NOT_REGISTERED`
will be sent.