Fixed Policy endpoints validation
For regular policies there was a wrong upper bound for endpoints
which allows for each endpoint to have a list with more than 3 items.
This issue was fixed for external policies.

The second part of issue is that PM does not check endpoints for
containing default group. This check was added for all policies flows.