Fix message param value replacing in HMI response
The root cause of this defect is that there was a wrong logic
in ResponseFromHMI for replacing "info" param with value of
"message" param. This param could be in msg_params section of
SmartObject in case of standart response and in params section
in case of erroneous response.

In this commit:
- Added GetMessageParamValue() function with correct logic for
reading "message" param from provided SmartObject
- Existing logic in ctor was replaced with function call