Add new functions for timer director for correct wait time calculation
Added GetSecondsToNearestTimeout() function to return time to the nearest
requests timeout or default timeout if no requests in queue. This is a
correct time to wait before checking timeout.

Added ResetTimer() function in TimerThreadDelegate and in TimeDoctor to
reset timer in related thread. This is used to notify thread to recalculate
time to wait until nearest request timeout.