Add TouchType CANCEL
During operation with a SDL application supporting video streaming,
a native dialog such as incoming-call may interrupt. If touch events
before and after a screen interruption are evaluated as one gesture,
an unexpected gesture may be detected. By notifying CANCEL, HMI can
cancel gesture recognitions of a SDL application.

This is for discussion on proposal SDL-0049.