Fix ButtonName INVALID_DATA response on any request
There was a problem with checking is button name present in HMI
capabilities, which SDL receives in RC.GetCapabilities response from
HMI. This capabilities is stored as enum values in smart object, however
in ButtonRequest button name param comes as string and SDL tries to find
this string in enum array which always finishes with false result.

To fix this issue was added map with pairs of string values and their
equivalent id values. This helps SDL to search button by its id which
exactly stored in HMI capabilities smart object.