Fix DISALLOWED response when app has specifically defined policies
This problem mostly related to RC functionality. When SDL receives
any RC request, it checks is sending app have REMOTE_CONTROL HMI type.
If app does not have such HMI type, SDL disallows such RPC for it.

The problem was in CacheManager::GetHMITypes() function, which
returns appHMIType array even if it was not initialized(param is not present
in policy) so SDL tries to find HMI type in an empty array and that always
evaluates to false so SDL decides that current app is not RC app. If appHMITypes
is not specified in policies, SDL should search HMI type in array of HMI types,
provided by application on its registration.

This check works correctly if application has default policies.

To fix that problem there was added check in GetHMITypes() to avoid
returning of uninitialized array.