Fix PTU applying for default app policy section
There was a problem with applying changes for applications
which is registered with "default" policies and this
policy group was updated after PTU. In this case permissions
for already registered applications which is using these groups
still have permissions as before update.
Also default policy section was not updated after PTU with changes
in this section.

This issue is reproduced on EXTERNAL_PROPRIETARY flow only.

To fix this issue there was removed code in CacheManager, which
incorrectly assigns default policies to apps with "default" policies.

Also there was a redundant code because default policies is unwrapped in
PTU before its applying, so all specific application policies is already
have actual new default policy permissions. In this case it is correct to
assign to every app his own policies from PTU.

Also there was updated logic in ProcessAppPolicyCheckResults() to perform
all needed actions once per app, because its possible that results could
contain sever results for one app_id.