Fix SendLocation response for not applicable resultCodes from HMI
Currently SDL transfers from HMI to Mobile any result code which
came in response. Some RPCs in MOBILE_API have own fixed list of
valid result codes for this RPC. If SDL receives some of result
code which is not present in this fixed list, SDL should behave
the same way as in case of "any unknown issue" and respond with
GENERIC_ERROR.

In this fix was added response message validation check after
applying schema and respond with GENERIC_ERROR if response
message is invalid due to some reason. This fix affects all RPC
responses.