fix: make sure to send OnAppRegistered notification after resumption
If HMI responds with OnAppActivated notification really quickly after
OnAppRegistered notification from SDL Core, it was possble that
the resumption ran after the app was activated, hence the HMI level
was reset to default value of the policy (which is NONE).

This change will make sure that OnAppActivated is always triggered
after resumption.