Fix application start streaming retry sequence
There was a problem with retry sequence which is
triggered when user rejecets/ignores prompt regarding
starting audio streaming on HMI. The problem is when
user clicks Cancel then SDL sends new request in the
sequence and not removes old one from queue. As a result
we have two identical requests and each of them could
trigger next iteration of retry sequence at the same time.
As a result, next try could be started earlier than expected.

For example if streaming retry sequence has 3 retries
with 1 second timeout:
- MOB -> SDL Request to start audio streaming
- SDL -> HMI Prompt user to start streaming for active app (request 1)
- User clicks "No" during 1 sec
- HMI -> SDL User has clicked No (request 1 still in queue!)
- SDL starts streaming retry sequence
- SDL -> HMI First retry and waiting for result during 1 sec (request 2)
- No answer from user during 1 sec
- Tiemout from request 1: trird retry
- Tiemout from request 2: retry limit reached. Streaming not allowed

In this update was added request termination in case user rejects
or accepts the request.