Fixed: app does not activate when policies are disabled.
Removed POLICY_LIB_CHECK_VOID() from PolicyHandler::OnActivateApp as
redundant, because the case with disabled policy library is already
handled inside that function. Also, it works incorrectly for the case
when the policy is disabled - this macro breaks the function at the
beginning so applications could not be activated when policy library is
disabled.