fix: process SDL.ActivateApp after resumption is triggered
If HMI sends SDL.ActivateApp request really quickly after
OnAppRegistered notification is sent from SDL Core, it was possble
that the resumption was triggered after the app was activated,
hence its HMI level was reset to default value of the policy
(which is NONE).

This change will make sure that SDL.ActivateApp is processed after
application's setup in RegisterAppInterfaceRequest::
SendRegisterAppInterfaceResponseToMobile() is completed, including
starting resumption.