Fixed RPC parameters permission calculation according to requirements
There was a problem when application has several consented functional
groups with the same RPC included. In case when some of functional
groups have RPC with all parameters disallowed option, SDL ignores all
allowed parameters for this RPC in other functional groups.

According to AppLink Policies Manager specification SDL should perform
a logical OR amongst all of the possible allowed permissions scenarios
for the RPC defined by each of the functional groups.

Current logic was updated to fit this requirement.