Set initialize state for the Interger type
The default constructor of Interger class assigns the min value
for the wrapped variable, hoowever it keeps the initialize state us
`Uninitialized`. So in case when latter user apply such operations as
`++`, `+=` the value will be changed but the state is not.

The commit adds ability to change initialize state for the class
after mentioned operations. So user does not have to call additionaly
`initialize` method.