Fix for state_controller unit tests.
Due to changes in a logic for PhoneCallHmiState,
unit tests have to be updated as well.
ApplicationType enum is removed since it is not actual anymore.
As a result, PrepareHMIStateResults functions are modified according to
new application types and updated expected HMI states.

Tests for NaviStreaming HMI state are enabled.
NaviStreamingHmiState is used for these tests instead of VideoStreamingHmiState.

New tests for different application types are added.