Fix moduleType saving into DB in EXTERNAL_PROPRIETARY
There was a missed logic of saving of moduleType parameter
of application into the database. SDL does not save this
parameter into the database when application get default
policy permissions. As a result, on the next ignition circle
SDL loads nothing from database for this application as
nothing was saved. As a result SDL rejects all RC related
requests due to wrongly set moduleType parameter.

In this commit was added missing logic of saving this
parameter into the database.