Fix non-media applications activation
There was a problem that SDL's state controller
does not take into account that application could provide
guarantees that it will not use audio/video source when it
is going to be activated. That's why SDL does not allow
non-media applications to be activated when some temporary
state (embedded navi or audio soirce) is currently active.
Was analyzed the following test cases:
- Media app (any hmi type) could not be activated during
AUDIO_SOURCE is active
- Navi or projection app could not be activated during
EMBEDDED_NAVI is active
- Navi or projection non-media apps could be activated during
AUDIO_SOURCE is active as they are using separate audio channels
for a voice prompts and does not affect embedded audio playing
- Media apps (except navi and projection) could be activated
during EMBEDDED_NAVI is active as they are using separate audio
channels