Open Menu RPC
Current commit contains brand new RPC "ShowAppMenu".
The built-in menu button shows the list of menu entries added
with AddCommand. But since this button is hidden for projection apps,
the menu is not accessible anymore. So the application should
be allowed to send a request to show the built-in menu view.

ShowAppMenu RPC allows projection application to send a request to open
the built-in app menu or any previously added sub menu
by a corresponding id.