fix: remove shared pointer of Application from HmiState (#2225)
* fix: remove shared pointer of Application from HmiState

This is to fix memory leak issue of ApplicationImpl instances
caused by cyclic reference of shared pointers.
Implementation in HmiState class is reverted to use app_id
instead of shared pointer.

* Reflect review comments

- store hmi_app_id rather than app_id
- use app_mngr_.application_by_hmi_app() to retrieve the
  shared pointer

Also, the old methods with app_id argument are marked as
deprecated again.

* Update mock functions in state controller test

This is to align the changes in hmi_state.cc.

* fix: application_manager_test crashes on Ubuntu 14.04

This only affects unit tests.

* Add hmi app id in the HmiState output

Reflecting review comment.

* fix: taking care ApplicationSharedPtr being null in hmi_state.cc

* fix: help_prompt_manager_test failures

- Remove outdated, unnecessary state creation
- Fix the order of destructions of member instances

* fix: state_controller_test failures

The test code was not correct. Actually RegularHmiState() method
of the mock class was never called, but because the instance wasn't
properly disposed the expectation was never verified.
The solution is to simply remove the inappropriate expectations.