SDLCORE-534: Changed behavior of destruction for thread and delegate thread
SDLCORE-534

We used aggregation in Thread class and we dont have contorol over
delegate life cycle, so logically need first to delete container and then delete delegate.
we are calling join() -> stop() Inside destuctor of Thread
we are using delegate_->exitThreadMain(); Inside the stop()

Shutdown () was removed because it is redundant. Destructor Thread will make this.