Fix SSL handshake for a secondary connection
There was found a several places in SDL code where
SSL handshake was working wrong when secured service
was started from a secondary connection like WIFI.
All these places were updated to consider primary
connection as a source for an SSL context and other
attributes for case when the current connection is
secondary.
This is a part which was not considered during
initial implementation of the secondary transport
feature.