Fix revoked VD items subscription after PTU
There was found an issue that SDL is still keeping
subscriptions for a custom vehicle data items even
for case when they was revoked by PTU. By that
reason, SDL was crashing with DCHECK on attempt to
unsubscribe from not existing custom data.
Was added missing logic of internal unsubscribe in
case if some custom vehicle data was revoked by
policies.