Synchronize backup and reset policy functions
This is an issue specific for EXTERNAL_PROPRIETARY flow
only. This issue is exposed by PolicyHandlerTest unit test:
ResetPolicyTable_WithPreloadedFile_ExpectPolicyTableReset
which sometimes is failing.
The main problem was that database reset(drop) and backup
can be done simultaneously in different threads which were
not synchronized. In that case, database drop often may be
failed because another thread is trying to write something
into the DB table which is going to be dropped. This causes
above unit test to fail sometimes.
To make that unit test and policy component more stable,
these two threads were synchronized by additing a waiting
function, so one thread will be able to wait for another
thread to finish its own staff.