Fix crash in WebsocketSession on start.
A crash occurs because we use an overload of
boost::beast::websocket::stream::write without error_code.
After the error occurs we add an error message to the logger
and continue working.

In SendMessage_UnpreparedConnection_WithoutFall test we use sleep because we can't ability to catch processed event.
This issue is that LoopThreadDelegate requires a WebsocketSession to work.
WebsocketSession init LoopThreadDelegate with 'this', so we have no way to simulate it without additional installers.

In future, to avoid sleep, we can:
Create a wrapper class for the socket, then provide the socket layout for the WebsocketSession in unit tests and wrap the following interfaces:
async_read
write down
get_executor
async_accept