Fix OnSystemRequest(LOCK_SCREEN_ICON_URL) sending
There was noticed an issue that sometimes Policy Manager
component is returning an empty url for lock screen icon.
The root cause of that is the data races between thread
which is sending OnSystemRequest notification and backupper
thread which overwrites data in the database. As these two
threads may have an intersection, there could be a possible
situation that backuper thread erased data from `endpoints`
table and another thread is using SQL query to select the
data from the same table. As a result, second thread receives
an empty result as the data was erased and the new one was not
inserted yet by backupper thread. After that, SDL just does
not send OnSystemRequest as the URL is empty.
To fix that issue, policy component has been updated to use
internal policy cache instead of direct access to policy
database each time to get URL. This will guarantee that
a sending thread will get an up-to-date infromation
required for OnSystemRequest notification.
Also, code cleanup has been done in order to remove all
functions which become unused after this fix.