Move logic for unsubscription from way points
There is a need in move this block of code, because deletion of data
about subscriptions before saving application data for resumption will
cause loss of data about subscription to way points. That's why logic for unsubsription
from way points needs to be located after code block, related to saving of application data for resumption.